requirements.txt
cloudinary==1.36.0
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.6
django-allauth==0.57.0
django-autoslug==1.9.9
django-cleanup==8.0.0
django-cloudinary-storage==0.3.0
django-htmx==1.16.0
django-infinite-scroll==0.1.7
django-mobile==0.7.0
Django-Mobileesp==1.0.1
django-rest-auth==0.9.5
django-stdimage==6.0.1
django-user-agents==0.4.0
django-user-unique-email==0.1.1
Django-Verify-Email==2.0.3
fontawesomefree==6.4.0
fonttools==4.39.4
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.8
psycopg2-binary==2.9.8
Pyrebase4==4.7.1
python-dotenv==1.0.0
regex==2023.10.3
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
user-agents==2.2.0
validators==0.22.0
whitenoise==6.5.0
